+++ spi-impl/series/series-dao/src/main/java/org/n52/io/extension/resulttime/ResultTimeRepository.java
@@ -7,8 +7,8 @@ import org.hibernate.Hibernate;
 import org.hibernate.Session;
 import org.n52.io.request.IoParameters;
 import org.n52.series.db.DataAccessException;
-import org.n52.series.db.SessionAwareRepository;
 import org.n52.series.db.beans.DatasetEntity;
+import org.n52.series.db.da.SessionAwareRepository;
 import org.n52.series.db.dao.DatasetDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
