+++ spi-impl/series/series-dao/src/main/java/org/n52/series/db/da/SessionAwareRepository.java
@@ -26,7 +26,7 @@
  * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
  * for more details.
  */
-package org.n52.series.db;
+package org.n52.series.db.da;
 
 import java.util.HashMap;
 import java.util.List;
