+++ spi-impl/series/series-dao/src/main/java/org/n52/series/db/dao/DatasetDao.java
@@ -134,7 +134,6 @@ public class DatasetDao<T extends DatasetEntity> extends AbstractDao<T> {
         return (List<T>) criteria.list();
     }
 
-
     @Override
     protected Class<T> getEntityClass() {
         return entityType;
